+++ src/com/ATP/ATPvehicle.java
@@ -2,25 +2,25 @@ package com.ATP;
 public class ATPvehicle {
 	int vehicleID;
 	int agentID;
-	int Vel;
+	double Vel;
 	double Eff;
-	
-	public ATPvehicle(int vel, double eff, int id){
+
+	public ATPvehicle(double vel, double eff, int id){
 		this.agentID = -1;
 		this.Vel = vel;
 		this.Eff = eff;
 		this.vehicleID = id;
 	}
-	
+
 	@Override
 	public int hashCode(){
 		int hash = 7;
 		hash = 31 * hash + this.vehicleID;
 		hash = 31 * hash + this.agentID;
-		hash = 31 * hash + this.Vel;
+		hash = 31 * hash + (int)this.Vel;
 		return hash;
 	}
-	
+
 	@Override
 	public boolean equals(Object other){
 		if (other == null)
@@ -31,48 +31,48 @@ public class ATPvehicle {
 				(this.Eff == vehicle.Eff) &&
 				(this.Vel == vehicle.Vel));
 	}
-	
+
 	@Override
 	public Object clone(){
 		ATPvehicle car = new ATPvehicle(Vel, Eff, vehicleID);
 		car.agentID = this.agentID;
 		return car;
 	}
-	
+
 	public boolean isTaken()
 	{
 		return (this.agentID > -1);
 	}
-	
+
 	public int getVehicleId()
 	{
 		return this.vehicleID;
 	}
-	
-	public int getVel()
+
+	public double getVel()
 	{
 		return this.Vel;
 	}
-	
+
 	public double speedFlooded()
 	{
-		//warning: if this.Eff == 0 then this will return 0!!! 
-		return this.Vel*this.Eff; 
+		//warning: if this.Eff == 0 then this will return 0!!!
+		return this.Vel*this.Eff;
 	}
-	
-	public int speedUnflooded()
+
+	public double speedUnflooded()
 	{
 		return this.Vel;
 	}
-	
+
 	public void setAgent(int id){
 		this.agentID = id;
 	}
-	
+
 	public int getAgent(){
 		return this.agentID;
 	}
-	
+
 	public double getEff()
 	{
 		return this.Eff;
