+++ src/com/ATP/Greedy.java
@@ -8,9 +8,9 @@ import com.waldura.tw.DijkstraEngine;
 
 
 public class Greedy extends Agent {
-	
+
 	private Vector<Integer> moves;
-	
+
 	public Greedy(int id, int initial, int goalv){
 		this.ID = id;
 		this.goal = goalv;
@@ -22,28 +22,28 @@ public class Greedy extends Agent {
 	public String toString(){
 		return "Greedy("+this.ID+", "+this.goal+")";
 	}
-	
+
 	public boolean reachedGoal(){
 		return (this.goal == this.position);
 	}
 
 	@Override
 	public ATPmove nextMove(ATPstate state) {
-		
+
 		if(moves == null){
 			DijkstraEngine.instance().execute(City.valueOf(state.getAgentPosition(this.ID)), City.valueOf(this.goal));
 			Vector<Integer> l = new Vector<Integer>();
-			 
+
 			 for (City city = City.valueOf(this.goal); city != null; city = DijkstraEngine.instance().getPredecessor(city))
 			 {
 			     l.add(city.getName());
 			 }
-			
+
 			 Collections.reverse(l);
 			 this.moves = l;
 			 this.moves.remove(0);
 		}
-		
+
 		if(this.goal != this.position){
 			boolean isFlooded;
 			int car = this.vehicleId;
@@ -59,22 +59,22 @@ public class Greedy extends Agent {
 			} catch (Exception e) {
 				e.printStackTrace();
 			}
-			
+
 			if (car== -1) return null;
 			return new ATPmove(moves.remove(0).intValue(), car);
 		} else return null;
 	}
-	
+
 	private boolean checkIfNextEdgeIsFlooded(int v, int u){
 		return ATPgraph.instance().getEdge(v, u).isFlooded();
 	}
-	
+
 	private int FindRegularVehicle(int pos, ATPstate state)
 	{
 		//we get in to this procedure only if this.vehicle == -1
 		Iterator<ATPvehicle> it = state.iterateVehicleAt(pos);
 		int bestVehicleId = -1;
-		int bestSpeed = 0;
+		double bestSpeed = 0;
 		while (it.hasNext())
 		{
 			ATPvehicle temp = it.next();
@@ -85,7 +85,7 @@ public class Greedy extends Agent {
 		}
 		return bestVehicleId;
 	}
-	
+
 	private int FindFloatingVehicle(int pos, ATPstate state){
 		Iterator<ATPvehicle> it = state.iterateVehicleAt(pos);
 		ATPvehicle best = null;
